TooActivationValidator
